Crate prisma_fmt

Functions

code_actions

This API is modelled on an LSP code action request. Input and output are both JSON, the request being a CodeActionParams object and the response being a list of CodeActionOrCommand objects.

format

The two parameters are:

get_config

This is the same command as get_config()

get_dmmf

lint

native_types

offset_to_position

Converts a byte offset to an LSP position, if the given offset does not overflow the document.

preview_features

referential_actions

text_document_completion

The API is modelled on an LSP completion request. Input and output are both JSON, the request being a CompletionParams object and the response being a CompletionList object.