Crate prisma_fmt

Functions

code_actions

This API is modelled on an LSP code action request. Input and output are both JSON, the request being a CodeActionParams object and the response being a list of CodeActionOrCommand objects.

format

The two parameters are:

get_config

This is the same command as get_config()

get_dmmf

This is the same command as get_dmmf()

lint

native_types

offset_to_position

Converts a byte offset to an LSP position, if the given offset does not overflow the document.

preview_features

referential_actions

text_document_completion

The API is modelled on an LSP completion request. Input and output are both JSON, the request being a CompletionParams object and the response being a CompletionList object.

validate

Function that throws a human-friendly error message when the schema is invalid, following the JSON formatting historically used by the Query Engine’s user_facing_errors::common::SchemaParserError. When the schema is valid, nothing happens. When the schema is invalid, the function displays a human-friendly error message indicating the schema lines where the errors lie and the total error count, e.g.: